requirements.txt
bcrypt==4.1.3
blinker==1.8.2
bson==0.5.10
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-mongoengine==1.0.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.24.1
PyJWT==2.8.0
pymongo==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
Werkzeug==3.0.3
WTForms==3.1.2
